+++ src/components/shared/Routers.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import { NoticeTree, Trees, MyPage, MessageSender, MessageDetail, MessageBox } from '@/pages';
-import TreeAddForm from '../features/Trees/TreeAddForm';
+import { NoticeTree, Forest, MyPage, MessageSender, MessageDetail, MessageBox } from '@/pages';
+import TreeAddForm from '../features/Forest/TreeAddForm';
 import ProfileEdit from '@/pages/MyPage/ProfileEdit';
 import Login from '@/pages/Login';
 
@@ -10,7 +10,7 @@ const Routers = () => {
 		<Routes>
 			<Route path="/" element={<NoticeTree />} />
 			<Route path="/login" element={<Login />} />
-			<Route path="/forest" element={<Trees />} />
+			<Route path="/forest" element={<Forest />} />
 			<Route path="/forest/edit" element={<TreeAddForm />} />
 			<Route path="/forest/edit/:treeId" element={<TreeAddForm />} />
 			<Route path="/mypage" element={<MyPage />} />
