+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 export { default as MessageBox } from './MessageBox';
 export { default as NoticeTree } from './NoticeTree';
-export { default as Trees } from './Trees';
+export { default as Forest } from './Forest';
 export { default as MessageSender } from './MessageSender';
 export { default as MessageDetail } from './MessageDetail/MessageDetail';
 export { default as MyPage } from './MyPage/MyPage';
